requirements.txt
# GMPFuzz - Python dependencies
ruamel.yaml>=0.17.0
click>=8.0
requests>=2.28.0
plotext>=5.0
autopep8>=2.0
